#36 自分が作成した問題のみを編集削除できる

+++ app/Http/Controllers/Santaku/DeleteController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Santaku;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Santaku;
+use App\Services\SantakuService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -14,9 +16,17 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, SantakuService $santakuService)
     {
         $santakuId = (int) $request->route('santakuId');
+        if (!$santakuService->checkOwnMondai($request->user()->id, $santakuId)) {
+            return redirect()
+                ->route('list')
+                ->with('feedback.success', "他のユーザーの問題は削除出来ません");
+
+            throw new AccessDeniedHttpException();
+        }
+
         $santaku = Santaku::where('id', $santakuId)->firstOrFail();
         $santaku->delete();
         return redirect()
